Moving siren entity interactions to separate behavior, adding a couple helper methods so that the activities list doesn't need to import Rels.
