+++ test/d2l-quick-eval/d2l-quick-eval-siren-helper-behavior.js
@@ -1,0 +1,296 @@
+import SirenParse from 'siren-parser';
+
+(function() {
+	var component;
+
+	suite('d2l-quick-eval-siren-helper-behavior', function() {
+		setup(function() {
+			component = fixture('basic');
+		});
+		suite('performSirenAction', function() {
+			test('when calling perform siren action with no query params and no fields, the fields are empty', () => {
+
+				const action = {
+					href : 'http://127.0.0.1/',
+					name: 'apply',
+					type: 'application/x-www-form-urlencoded',
+					method: 'GET'
+				};
+
+				const stub = sinon.stub(component, 'performSirenAction', function(passedAction) {
+					assert.deepEqual(action, passedAction);
+				});
+
+				component._performSirenActionWithQueryParams(action);
+				sinon.assert.calledWith(stub, action);
+			});
+			test('when calling perform siren action with no query params, the fields are not modified', () => {
+
+				const action = {
+					href : 'http://127.0.0.1/',
+					name: 'apply',
+					type: 'application/x-www-form-urlencoded',
+					method: 'GET',
+					fields : [
+						{
+							type: 'hidden',
+							name : 'existingField',
+							value: 'existingValue'
+						}]
+				};
+
+				const stub = sinon.stub(component, 'performSirenAction', function(passedAction) {
+					assert.deepEqual(action, passedAction);
+				});
+
+				component._performSirenActionWithQueryParams(action);
+				sinon.assert.calledWith(stub, action);
+			});
+			test('when calling perform siren action with no query params and custom params, fields contain custom params', () => {
+
+				const action = {
+					href : 'http://127.0.0.1/',
+					name: 'apply',
+					type: 'application/x-www-form-urlencoded',
+					method: 'GET'
+				};
+
+				const customParams = { customParam1: 'custom', customParam2: 'custom2' };
+				sinon.stub(component, 'performSirenAction', function(passedAction) {
+					const fields = passedAction.fields;
+					assert.equal(Object.keys(customParams).length, fields.length);
+
+					Object.keys(customParams).forEach(function(p) {
+						assert.isTrue(fields.some(function(elm) { return elm.name === p && elm.value === customParams[p]; }));
+					});
+				});
+
+				component._performSirenActionWithQueryParams(action, customParams);
+			});
+			test('when calling perform siren action with query params, the query params are added as fields', () => {
+
+				const action = {
+					href : 'http://127.0.0.1?testname=testvalue&anothertestname=anothertestvalue',
+					name: 'apply',
+					type: 'application/x-www-form-urlencoded',
+					method: 'GET',
+					fields : [
+						{
+							type: 'hidden',
+							name : 'existingField',
+							value: 'existingValue'
+						}]
+				};
+
+				const expectedAction = {
+					href : 'http://127.0.0.1?testname=testvalue&anothertestname=anothertestvalue',
+					name: 'apply',
+					type: 'application/x-www-form-urlencoded',
+					method: 'GET',
+					fields : [
+						{
+							type: 'hidden',
+							name : 'existingField',
+							value: 'existingValue'
+						},
+						{
+							type: 'hidden',
+							name : 'testname',
+							value: 'testvalue'
+						},
+						{
+							type: 'hidden',
+							name : 'anothertestname',
+							value: 'anothertestvalue'
+						}]
+				};
+
+				const stub = sinon.stub(component, 'performSirenAction', function(passedAction) {
+					assert.deepEqual(expectedAction, passedAction);
+				});
+
+				component._performSirenActionWithQueryParams(action);
+				sinon.assert.calledWith(stub, action);
+			});
+			test('when calling perform siren action with query params and custom params, fields contain query params and custom params', () => {
+
+				const defaultParams = [
+					{
+						type: 'hidden',
+						name : 'testname',
+						value: 'testvalue'
+					},
+					{
+						type: 'hidden',
+						name : 'anothertestname',
+						value: 'anothertestvalue'
+					}
+				];
+
+				const action = {
+					href : 'http://127.0.0.1/',
+					name: 'apply',
+					type: 'application/x-www-form-urlencoded',
+					method: 'GET',
+					fields : defaultParams
+				};
+
+				const customParams = { customParam1: 'custom', customParam2: 'custom2' };
+				sinon.stub(component, 'performSirenAction', function(passedAction) {
+					const fields = passedAction.fields;
+					assert.equal(4, fields.length);
+
+					assert.isTrue(fields.some(function(elm) { return elm.name === 'testname' && elm.value === 'testvalue'; }));
+					assert.isTrue(fields.some(function(elm) { return elm.name === 'anothertestname' && elm.value === 'anothertestvalue'; }));
+
+					Object.keys(customParams).forEach(function(p) {
+						assert.isTrue(fields.some(function(elm) { return elm.name === p && elm.value === customParams[p]; }));
+					});
+				});
+
+				component._performSirenActionWithQueryParams(action, customParams);
+			});
+		});
+		suite('parsing url (_getExtraParams)', function() {
+			test('when parsing url for sort and filter params and url is null, return empty array', () => {
+
+				var params = component._getExtraParams('');
+				assert.equal(params.length, 0);
+			});
+			test('when parsing url for sort and filter params and url is null, return empty array', () => {
+
+				var params = component._getExtraParams(null);
+				assert.equal(params.length, 0);
+			});
+			test('when parsing url for sort and filter params, if they are both present, return array with correct values', () => {
+				const url = 'https://www.example.com/?pageSize=20&filter=96W3siU29ydCI6eyJJ&sort=Y3Rpb24iOjB9';
+
+				var params = component._getExtraParams(url);
+				assert.equal(params.length, 2);
+
+				const expectedParams = [
+					{
+						name: 'filter',
+						value: '96W3siU29ydCI6eyJJ'
+					},
+					{
+						name: 'sort',
+						value: 'Y3Rpb24iOjB9'
+					}
+				];
+				assert.deepEqual(params, expectedParams);
+			});
+			test('when parsing url for sort and filter params, if only sort is present, return array with correct values', () => {
+				const url = 'https://www.example.com/?pageSize=20&sort=Y3Rpb24iOjB9';
+
+				var params = component._getExtraParams(url);
+				assert.equal(params.length, 1);
+
+				const expectedParams = [
+					{
+						name: 'sort',
+						value: 'Y3Rpb24iOjB9'
+					}
+				];
+				assert.deepEqual(params, expectedParams);
+			});
+			test('when parsing url for sort and filter params, if only filter is present, return array with correct values', () => {
+				const url = 'https://www.example.com/?pageSize=20&filter=96W3siU29ydCI6eyJJ';
+
+				var params = component._getExtraParams(url);
+				assert.equal(params.length, 1);
+
+				const expectedParams = [
+					{
+						name: 'filter',
+						value: '96W3siU29ydCI6eyJJ'
+					}
+				];
+				assert.deepEqual(params, expectedParams);
+			});
+		});
+		suite('creating evaluation link (_buildRelativeUri)', function() {
+			test('when creating the evaluation link, if there are no extra params, return original link', () => {
+				const url = '/d2l/lms/tool/mark.d2l?ou=122041&db=1004';
+				const params = [];
+
+				var evalLink = component._buildRelativeUri(url, params);
+				assert.equal(evalLink, url);
+			});
+			test('when creating the evaluation link, if there are extra params, return correct link', () => {
+				const url = '/d2l/lms/tool/mark.d2l?ou=122041&db=1004';
+				const params = [
+					{
+						name: 'filter',
+						value: '96W3siU29ydCI6eyJJ'
+					},
+					{
+						name: 'sort',
+						value: 'Y3Rpb24iOjB9'
+					}
+				];
+				const expectedEvalLink = url + '&filter=96W3siU29ydCI6eyJJ&sort=Y3Rpb24iOjB9';
+
+				var evalLink = component._buildRelativeUri(url, params);
+				assert.equal(evalLink, expectedEvalLink);
+			});
+			test('when creating the evaluation link, if there are extra params and url has no original params, return correct link', () => {
+				const url = '/d2l/lms/tool/122041/mark/1004/';
+				const params = [
+					{
+						name: 'filter',
+						value: '96W3siU29ydCI6eyJJ'
+					},
+					{
+						name: 'sort',
+						value: 'Y3Rpb24iOjB9'
+					}
+				];
+				const expectedEvalLink = url + '?filter=96W3siU29ydCI6eyJJ&sort=Y3Rpb24iOjB9';
+
+				var evalLink = component._buildRelativeUri(url, params);
+				assert.equal(evalLink, expectedEvalLink);
+			});
+		});
+		suite('_tryGetName', function() {
+			test('_tryGetName returns default value when subentity has class "default-name"', () => {
+				const nameRel = 'nameRel';
+				const expectedName = 'defaultValue';
+
+				const userEntity = {
+					'entities': [
+						{
+							'class': ['default-name'],
+							'rel': [ nameRel ],
+							'properties': {
+								'name': 'someName'
+							}
+						}
+					]
+				};
+
+				const name = component._tryGetName(SirenParse(userEntity), nameRel, expectedName);
+				assert.equal(name, expectedName);
+			});
+			test('_tryGetName returns expected name value name subentity is valid', () => {
+				const nameRel = 'nameRel';
+				const expectedName = 'expectedName';
+
+				const userEntity = {
+					'entities': [
+						{
+							'class': [],
+							'rel': [ nameRel ],
+							'properties': {
+								'name': expectedName
+							}
+						}
+					]
+				};
+
+				const name = component._tryGetName(SirenParse(userEntity), nameRel, 'defaultValue');
+				assert.equal(name, expectedName);
+			});
+		});
+	});
+})();
